+++ packages/backend/src/routers/projectRouter.ts
@@ -23,6 +23,7 @@ import {
     searchService,
     spaceService,
 } from '../services/services';
+import { integrationsRouter } from './integrations/intergrationsRouter';
 
 export const projectRouter = express.Router({ mergeParams: true });
 
@@ -610,3 +611,5 @@ projectRouter.delete(
         }
     },
 );
+
+projectRouter.use('/integrations', integrationsRouter);