+++ packages/common/src/index.ts
@@ -2,6 +2,10 @@ import moment from 'moment';
 import { v4 as uuidv4 } from 'uuid';
 import { Dashboard, DashboardBasicDetails } from './types/dashboard';
 import { convertAdditionalMetric } from './types/dbt';
+import {
+    DbtCloudIntegration,
+    DbtCloudMetadataResponseMetrics,
+} from './types/dbtCloud';
 import { Explore, SummaryExplore } from './types/explore';
 import {
     CompiledDimension,
@@ -68,6 +72,7 @@ export * from './templating/template';
 export * from './types/api';
 export * from './types/dashboard';
 export * from './types/dbt';
+export * from './types/dbtCloud';
 export * from './types/errors';
 export * from './types/explore';
 export * from './types/field';
@@ -87,6 +92,7 @@ export * from './types/table';
 export * from './types/timeFrames';
 export * from './types/user';
 export * from './utils/api';
+export { default as assertUnreachable } from './utils/assertUnreachable';
 export * from './utils/formatting';
 export * from './utils/timeFrames';
 
@@ -663,6 +669,8 @@ type ApiResults =
     | ProjectMemberProfile[]
     | SearchResults
     | Space
+    | DbtCloudMetadataResponseMetrics
+    | DbtCloudIntegration
     | ShareUrl;
 
 export type ApiResponse = {
@@ -788,11 +796,12 @@ export type BigqueryCredentials = Omit<
 
 export type CreateDatabricksCredentials = {
     type: WarehouseTypes.DATABRICKS;
-    serverHostName: string;
-    port: number;
+    catalog?: string;
+    // this supposed to be a `schema` but changing it will break for existing customers
     database: string;
-    personalAccessToken: string;
+    serverHostName: string;
     httpPath: string;
+    personalAccessToken: string;
 };
 
 export type DatabricksCredentials = Omit<
@@ -853,6 +862,7 @@ export type CreateSnowflakeCredentials = {
     threads?: number;
     clientSessionKeepAlive?: boolean;
     queryTag?: string;
+    accessUrl?: string;
 };
 
 export type SnowflakeCredentials = Omit<
@@ -1192,8 +1202,4 @@ export const deepEqual = (
             (!areObjects && val1 !== val2)
         );
     });
-};
-
-export const assertUnreachable = (_x: never): never => {
-    throw new Error("Didn't expect to get here");
 };