+++ packages/frontend/src/components/UnderlyingData/UnderlyingDataModalContent.tsx
@@ -61,6 +61,14 @@ const UnderlyingDataModalContent: FC<Props> = () => {
         [explore],
     );
 
+    const showUnderlyingValues: string[] | undefined = useMemo(() => {
+        return config?.meta !== undefined &&
+            isField(config?.meta?.item) &&
+            isMetric(config?.meta.item)
+            ? config?.meta.item.showUnderlyingValues
+            : undefined;
+    }, [config?.meta]);
+
     const metricQuery = useMemo<MetricQuery>(() => {
         if (!config) return defaultMetricQuery;
         const { meta, row, pivot, dimensions, value } = config;
@@ -170,10 +178,6 @@ const UnderlyingDataModalContent: FC<Props> = () => {
             },
         };
 
-        const showUnderlyingValues: string[] | undefined =
-            isField(meta?.item) && isMetric(meta.item)
-                ? meta.item.showUnderlyingValues
-                : undefined;
         const showUnderlyingTable: string | undefined = isField(meta?.item)
             ? meta.item.table
             : undefined;
@@ -196,7 +200,15 @@ const UnderlyingDataModalContent: FC<Props> = () => {
             dimensions: dimensionFields,
             filters: allFilters,
         };
-    }, [config, filters, tableName, allFields, allDimensions, joinedTables]);
+    }, [
+        config,
+        filters,
+        tableName,
+        allFields,
+        allDimensions,
+        joinedTables,
+        showUnderlyingValues,
+    ]);
 
     const fieldsMap: Record<string, Field> = useMemo(() => {
         const selectedDimensions = metricQuery.dimensions;
@@ -213,9 +225,16 @@ const UnderlyingDataModalContent: FC<Props> = () => {
     }, [explore, metricQuery]);
 
     const exploreFromHereUrl = useMemo(() => {
+        const showDimensions =
+            showUnderlyingValues !== undefined ? metricQuery.dimensions : [];
+
         const createSavedChartVersion: CreateSavedChartVersion = {
             tableName,
-            metricQuery,
+            metricQuery: {
+                ...metricQuery,
+                dimensions: showDimensions,
+                metrics: [],
+            },
             pivotConfig: undefined,
             tableConfig: {
                 columnOrder: [],
@@ -230,7 +249,7 @@ const UnderlyingDataModalContent: FC<Props> = () => {
             createSavedChartVersion,
         );
         return `${pathname}?${search}`;
-    }, [tableName, metricQuery, projectUuid]);
+    }, [tableName, metricQuery, projectUuid, showUnderlyingValues]);
 
     const { data: resultsData, isLoading } = useUnderlyingDataResults(
         tableName,