+++ packages/frontend/src/utils/fieldValidators.ts
@@ -25,6 +25,12 @@ export const startWithSlash: FieldValidator<string> = (fieldName) => (value) =>
         ? undefined
         : `${fieldName} should start with a "/"`;
 
+export const startWithHTTPSProtocol: FieldValidator<string> =
+    (fieldName) => (value) =>
+        !value || value.match(/^https:\/\/.*/)
+            ? undefined
+            : `${fieldName} should start with a "https://"`;
+
 export const isValidEmail: FieldValidator<string> = (fieldName) => (value) =>
     !value || validateEmail(value) ? undefined : `${fieldName} is not valid`;